Struct necsim_impls_no_std::cogs::active_lineage_sampler::alias::location::LocationAliasActiveLineageSampler

pub struct LocationAliasActiveLineageSampler<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: GloballyCoherentLineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: GillespieEventSampler<M, H, G, S, X, D, C, T, N>, I: ImmigrationEntry<M>> { /* private fields */ }

Implementations

impl<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: GloballyCoherentLineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: GillespieEventSampler<M, H, G, S, X, D, C, T, N>, I: ImmigrationEntry<M>> LocationAliasActiveLineageSampler<M, H, G, S, X, D, C, T, N, E, I>

pub fn init_with_store<'h, O: TrustedOriginSampler<'h, M, Habitat = H>>( origin_sampler: O, dispersal_sampler: &D, coalescence_sampler: &C, turnover_rate: &T, speciation_probability: &N, event_sampler: &E ) -> (S, Self)where H: 'h,

pub fn resume_with_store<'h, O: UntrustedOriginSampler<'h, M, Habitat = H>>( origin_sampler: O, dispersal_sampler: &D, coalescence_sampler: &C, turnover_rate: &T, speciation_probability: &N, event_sampler: &E, resume_time: NonNegativeF64 ) -> (S, Self, Vec<ExceptionalLineage>)where H: 'h,