Trait necsim_impls_no_std::cogs::event_sampler::gillespie::GillespieEventSampler

pub trait GillespieEventSampler<M: MathsCore, H: Habitat<M>, G: RngCore<M>, S: LineageStore<M, H>, X: EmigrationExit<M, H, G, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>>: EventSampler<M, H, G, S, X, D, C, T, N> {
    fn get_event_rate_at_location(
        &self,
        location: &Location,
        habitat: &H,
        lineage_store: &S,
        dispersal_sampler: &D,
        coalescence_sampler: &C,
        turnover_rate: &T,
        speciation_probability: &N
    ) -> NonNegativeF64 { ... }
}

Provided Methods

fn get_event_rate_at_location(
&self,
location: &Location,
habitat: &H,
lineage_store: &S,
dispersal_sampler: &D,
coalescence_sampler: &C,
turnover_rate: &T,
speciation_probability: &N
) -> NonNegativeF64

Pre: all lineages must be active in the lineage store