Struct necsim_impls_no_std::cogs::lineage_reference::in_memory::InMemoryLineageReference

#[repr(transparent)]
pub struct InMemoryLineageReference(_);

Implementations

impl InMemoryLineageReference

pub unsafe fn issue(reference: usize) -> Self

Safety

A new InMemoryLineageReference must only be issued from a usize to create a new and unique reference. It is not allowed to issue multiple InMemoryLineageReferences from the same usize.

In case where a temporary and unexposed clone of the InMemoryLineageReference is required, use [Backup::backup_unchecked].