Trait necsim_impls_no_std::cogs::active_lineage_sampler::singular::SingularActiveLineageSampler

pub trait SingularActiveLineageSampler<M: MathsCore, H: Habitat<M>, G: RngCore<M>, R: LineageReference<M, H>, S: LineageStore<M, H, R>, X: EmigrationExit<M, H, G, R, S>, D: DispersalSampler<M, H, G>, C: CoalescenceSampler<M, H, R, S>, T: TurnoverRate<M, H>, N: SpeciationProbability<M, H>, E: EventSampler<M, H, G, R, S, X, D, C, T, N>, I: ImmigrationEntry<M>>: ActiveLineageSampler<M, H, G, R, S, X, D, C, T, N, E, I> {
    fn replace_active_lineage(
        &mut self, 
        active_lineage: Option<Lineage>
    ) -> Option<Lineage>;
}

Required methods

fn replace_active_lineage(
&mut self,
active_lineage: Option<Lineage>
) -> Option<Lineage>